listings-position-B
listings-position-A from Chinatown NH is a Limited so I can not see it
full view unless I log it. The log in modal is appearing right after I
click on the listing so I can not test the Contact Agent button. Changed
to a listing that gives me this permission.